+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Version changes are pinned to SDK releases.
 
 ## [Unreleased]
 
+- client: Add in functionality to close a margin account, close an open orders account and close multiple open orders accounts. ([#93](https://github.com/zetamarkets/sdk/pull/93))
+- general: Add in functionality to accomodate for an extra order type post-only. ([#93](https://github.com/zetamarkets/sdk/pull/93))
+
 ## [0.11.0] 2022-02-02
 
 - risk: Change margin calculations to reflect new changes. See https://zetamarkets.gitbook.io/zeta/zeta-protocol/collateral-framework ([#85](https://github.com/zetamarkets/sdk/pull/85))