+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Version changes are pinned to SDK releases.
 ## [Unreleased]
 
 - client: Add in functionality to close a margin account, close an open orders account and close multiple open orders accounts. ([#93](https://github.com/zetamarkets/sdk/pull/93))
-- general: Add in functionality to accomodate for an extra order type post-only. ([#93](https://github.com/zetamarkets/sdk/pull/93))
+- general: Add in functionality to accomodate for extra order types (post-only & fill-or-kill). ([#93](https://github.com/zetamarkets/sdk/pull/93))
 
 ## [0.11.0] 2022-02-02